refactor(common): removes signal operators
keep isNonNull, rest ist rxjs duplicate I guess
