+++ libs/common/src/signal/empty.spec.ts
