+++ libs/common/src/signal/empty.ts
