+++ libs/common/src/signal/signalize.ts
@@ -1,3 +1,9 @@
 import { Signal, isSignal, signal } from '@angular/core';
 
-export const signalize = <T>(x: T | Signal<T>) => (isSignal(x) ? x : signal(x));
+/**
+ *
+ * @param source
+ * @returns the source when it is a signal, else creates a new signal with the source as its initial value.
+ */
+export const signalize = <T>(source: T | Signal<T>) =>
+  isSignal(source) ? source : signal(source);
